restructured frontend project folder
