+++ Frontend/src/components/general/BrowseTemplates.js
